requirements/prod.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
elasticsearch==7.8.0
idna==2.10
langdetect==1.0.8
lxml==4.5.1
PySocks==1.7.1
PyYAML==5.3.1
requests==2.24.0
schedule==0.6.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
stem==1.8.0
urllib3==1.25.9
xlrd==1.2.0
Flask==2.2.5
